shipit_code_coverage: Refactor artifacts- and github-related code in separate modules
