+++ src/shipit_code_coverage/tests/test_artifacts.py
@@ -2,7 +2,7 @@
 
 import os
 from shipit_code_coverage import utils
-from shipit_code_coverage.codecov import CodeCov
+from shipit_code_coverage.artifacts import ArtifactsHandler
 
 
 FILES = ['ccov-artifacts/%s' % f for f in [
@@ -22,8 +22,8 @@ def test_get_chunks():
     for f in FILES:
         open(f, 'w')
 
-    c = CodeCov(None, 'ccov-artifacts', '', '', '', '', None, None)
-    assert set(c.get_chunks()) == set([
+    a = ArtifactsHandler([], [])
+    assert set(a.get_chunks()) == set([
         'mochitest-1', 'mochitest-2', 'xpcshell-3', 'xpcshell-7',
         'cppunit', 'firefox-ui-functional-remote',
     ])
@@ -38,17 +38,17 @@ def test_get_coverage_artifacts():
     for f in FILES:
         open(f, 'w')
 
-    c = CodeCov(None, 'ccov-artifacts', '', '', '', '', None, None)
-    assert set(c.get_coverage_artifacts()) == set(FILES)
-    assert set(c.get_coverage_artifacts(suite='mochitest')) == set([
+    a = ArtifactsHandler([], [])
+    assert set(a.get_coverage_artifacts()) == set(FILES)
+    assert set(a.get_coverage_artifacts(suite='mochitest')) == set([
         'ccov-artifacts/windows_mochitest-1_code-coverage-jsvm.info',
         'ccov-artifacts/linux_mochitest-2_code-coverage-grcov.zip'
     ])
-    assert set(c.get_coverage_artifacts(chunk='xpcshell-7')) == set([
+    assert set(a.get_coverage_artifacts(chunk='xpcshell-7')) == set([
         'ccov-artifacts/windows_xpcshell-7_code-coverage-jsvm.info',
         'ccov-artifacts/linux_xpcshell-7_code-coverage-grcov.zip'
     ])
-    assert set(c.get_coverage_artifacts(chunk='cppunit')) == set([
+    assert set(a.get_coverage_artifacts(chunk='cppunit')) == set([
         'ccov-artifacts/windows_cppunit_code-coverage-grcov.zip'
     ])
